Support different backend for running inference. The server hosting the VSOC might not contain a GPU. This change adds the possibility to fetch the inference results from a remote HTTP server through an OpenAI inspired embeddings API.
